third_party/blink/web_tests/svg/dynamic-updates/SVGRadialGradientElement-svgdom-fx-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'fx' property of the SVGRadialGradientElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS radialGradientElement.fx.baseVal.valueAsString is "0%"
PASS radialGradientElement.fx.baseVal.valueAsString is "50%"
PASS successfullyParsed is true

TEST COMPLETE
